Remove 94 no-op width casts across 29 settled units
Two shapes, both restating a width the expression already has:

- Function-return casts. A call already delivers its declared return type
  in r3, so (u32)mainGetBit(...) / (u8)Obj_IsLoadingLocked() /
  (int)getCurGameText() / (u16)modelFileHeaderGetCullDistance(...) /
  (int)ObjGroup_ContainsObject(...) / (u8)baddie_canSeeTarget(...) are
  noise. Verified against the single repo-wide prototype for each callee
  (gamebits_api.h, sys/objects.h, textrender_api.h, model.h, obj_group.h,
  dll_00C4_tricky.h) and confirmed no TU here sees a differing width.

- Field casts restating the field's own declared type: (s8) on the s8
  rotationXByte / spawnYawByte placement bytes, (s16) on the s16
  contentsSpawnBitValue, (u16) on the u16 scaleTarget. A memory load
  already extends to the declared type.

Each candidate field required exactly one scalar declaration repo-wide
and no pointer/aggregate declaration anywhere. Every unit's .o is
byte-identical after the change (objcopy -R .comment, then
objdump -M gekko -s -t -r, diffed per file); ninja all_source and ninja
clean, main.dol OK.

+++ src/dlls/objects/239/239.c
@@ -1269,7 +1269,7 @@ void pushable_init(GameObject* obj, PushableObjectDef* setup) {
     state->unkB0 = *(int*)&setup->unk1C;
     state->scale = (f32) * &setup->scaleRaw / PUSHABLE_SCALE_DENOM;
     state->scale = state->scale * obj->anim.modelInstance->rootMotionScaleBase;
-    state->cullDistance = state->scale * (f32)(u16)modelFileHeaderGetCullDistance((ModelFileHeader*)*activeModelSlot) +
+    state->cullDistance = state->scale * (f32)modelFileHeaderGetCullDistance((ModelFileHeader*)*activeModelSlot) +
                           PUSHABLE_MIN_GROUND_CLEARANCE;
     {
         f32 z0 = PUSHABLE_ZERO;
